Stub out failing GraphQL test and old feature test
Since the frontend is written in totally React, it will be easiest to
setup those tests in Jest.

Also, the GraphQL API test I wrote is failing, but I don't know how to
make it pass (and also don't know how to get the functionality working
in the first place!).

I don't want to focus too much on GraphQL right now, since it might be a
bit overkill for the simple CRUD stuff we are doing for the uploader.

It will be very useful for the library feature though!