Use ReactRouter to implement nested routes
Switching over to a nested route structure similar to Rails Resourceful
routing.

This significantly simplifies a lot of the code.

Now, each "page" is represented in frontend/src/pages, and is referenced
in index.tsx at the top level.

Eventually we may want to pull in useContext in order to handle loading
the Zines on all the pages, but for now this is still way better.